replaced urls with reverse
